+++ src/components/UpupUploader/MethodSelector.tsx
@@ -1,40 +1,40 @@
 import type { Method } from 'types/Method'
-
-import React from 'react'
+import { MutableRefObject } from 'react'
 
 const MethodsSelector = ({
     setView,
     inputRef,
     methods,
-    isDocument,
 }: {
     setView: (view: string) => void
-    inputRef?: React.MutableRefObject<HTMLInputElement | null>
+    inputRef?: MutableRefObject<HTMLInputElement | null>
     methods: Method[]
-    isDocument: boolean
 }) => {
     return (
-        <div className="w-full h-full flex flex-col items-center justify-center gap-6 ">
+        <div className="w-full h-full flex flex-col items-center justify-center gap-6">
             <h1 className="md:text-2xl text-center dark:text-white">
-                Drop your {isDocument ? 'files' : 'images'} here,{' '}
+                Drop your files here,{' '}
                 <button
                     className="text-[#3782da] hover:underline"
                     onClick={() => inputRef && inputRef.current!.click()}
+                    type="button"
                 >
-                    browse {isDocument ? 'files' : 'images'}
+                    browse files
                 </button>{' '}
                 or import from:
             </h1>
             <div className="grid grid-cols-3 sm:grid-cols-4 md:grid-cols-6 grid-rows-2">
                 {methods.map(method => (
                     <button
+                        type="button"
                         key={method.id}
                         className="flex flex-col items-center justify-center gap-1 text-sm dark:hover:bg-[#282828] hover:bg-[#e9ecef] dark:active:bg-[#333] active:bg-[#dfe6f1] rounded-md p-2 px-4 transition-all duration-300 mb-4 disabled:opacity-30 disabled:pointer-events-none group relative"
                         disabled={
-                            !/INTERNAL|GOOGLE_DRIVE|ONE_DRIVE|LINK|CAMERA/.test(
-                                method.id,
-                            )
+                            !/INTERNAL|GOOGLE_DRIVE|LINK|CAMERA/.test(method.id)
                         }
+                        onKeyDown={e => {
+                            if (e.key === 'Enter') e.preventDefault()
+                        }}
                         onClick={() =>
                             method.id === 'INTERNAL'
                                 ? inputRef && inputRef.current!.click()