+++ src/types/BaseConfigs.ts
@@ -2,5 +2,8 @@ export type BaseConfigs = {
     toBeCompressed?: boolean
     multiple?: boolean
     onChange?: (files: File[]) => void
-    isDocument?: boolean
+    accept?: string
+    limit?: number
+    onFileClick?: (file: File) => void
+    mini?: boolean
 }